Permisos

+++ src/main/java/ar/edu/unlam/tallerweb1/controladores/UsuarioController.java
@@ -12,7 +12,8 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam; 
 import org.springframework.web.servlet.ModelAndView; 
 import ar.edu.unlam.tallerweb1.modelo.RolModel; 
-import ar.edu.unlam.tallerweb1.modelo.UsuarioModel; 
+import ar.edu.unlam.tallerweb1.modelo.UsuarioModel;
+import ar.edu.unlam.tallerweb1.modelo.enums.Rol;
 import ar.edu.unlam.tallerweb1.modelo.form.FormularioAgregarUsuario; 
 import ar.edu.unlam.tallerweb1.servicios.RolService; 
 import ar.edu.unlam.tallerweb1.servicios.UsuarioRolService; 
@@ -28,23 +29,25 @@ import ar.edu.unlam.tallerweb1.servicios.UsuarioService;
   
 	@RequestMapping("/usuarios") 
 	public ModelAndView usuarios(HttpServletRequest request) {
-  
-		String rol = request.getSession().getAttribute("ROL") != null ? (String)request.getSession().getAttribute("ROL") : "";
-  
-		if (!rol.equals("Admin")) 
-			return new ModelAndView ("redirect:/login");
+		Long rol = (Long)request.getSession().getAttribute("ROL");
+		if (rol != Rol.ADMIN.getId()) 
+			return new ModelAndView ("redirect:/logout");
   
 		ModelMap modelo = new ModelMap();
 		modelo.put("titulo", "Lista de Usuarios"); 
 		modelo.put("usuarios", usuarioService.listarUsuarios()); 
 		modelo.put("nombreUsuario", request.getSession().getAttribute("NOMBRE"));
+		modelo.put("rol", request.getSession().getAttribute("ROL"));
   
 		return new ModelAndView("usuarios", modelo); 
 	}
   
 	@RequestMapping(path = "/agregarUsuario", method = RequestMethod.POST) 
 	public ModelAndView agregarUsuario(HttpServletRequest request) { 
-  
+		Long rol = (Long)request.getSession().getAttribute("ROL");
+		if (rol != Rol.ADMIN.getId()) 
+			return new ModelAndView ("redirect:/logout");
+		
 		FormularioAgregarUsuario formulario = new FormularioAgregarUsuario();
 		List<RolModel> listDeRoles = rolService.listarRolUsuario();
 	  
@@ -52,19 +55,25 @@ import ar.edu.unlam.tallerweb1.servicios.UsuarioService;
 		modelo.put("titulo", "Agregar Usuario");
 		modelo.put("formularioAgregarUsuario", formulario);
 		modelo.put("listaDeRoles", listDeRoles); modelo.put("nombreUsuario", request.getSession().getAttribute("NOMBRE"));
-  
+		modelo.put("rol", request.getSession().getAttribute("ROL"));
+
 		return new ModelAndView("agregarUsuario", modelo); 
 	}
   
 	@RequestMapping(path = "/validarRegistroUsuario", method = RequestMethod.POST) 
 	public ModelAndView validarRegistroUsuario(@ModelAttribute("formularioAgregarUsuario") FormularioAgregarUsuario formularioAgregarUsuario, HttpServletRequest request) { 
+		Long rol = (Long)request.getSession().getAttribute("ROL");
+		if (rol != Rol.ADMIN.getId()) 
+			return new ModelAndView ("redirect:/logout");
+		
 		UsuarioModel usuario = formularioAgregarUsuario.getUsuario();
 		List<RolModel> listDeRoles = rolService.listarRolUsuario();
 		
 		ModelMap modelo = new ModelMap();
 		modelo.put("titulo", "Agregar Usuario"); modelo.put("listaDeRoles", listDeRoles); 
 		modelo.put("nombreUsuario", request.getSession().getAttribute("NOMBRE"));
-  
+		modelo.put("rol", request.getSession().getAttribute("ROL"));
+
 		if (usuarioService.validarRegistroUsuario(formularioAgregarUsuario) == false) { 
 			modelo.put("errorValidacion", "El nombre de usuario o email ya existe, contacte al administrador"); 
 			return new ModelAndView("agregarUsuario", modelo);
@@ -77,20 +86,29 @@ import ar.edu.unlam.tallerweb1.servicios.UsuarioService;
 	  
 	@RequestMapping("/editarUsuario") 
 	public ModelAndView editarUsuario(@RequestParam("id") Long id, HttpServletRequest request) {
-  
+		Long rol = (Long)request.getSession().getAttribute("ROL");
+		if (rol != Rol.ADMIN.getId()) 
+			return new ModelAndView ("redirect:/logout");
+		
 		UsuarioModel usuario = usuarioService.buscarUsuarioPorId(id);
 		ModelMap modelo = new ModelMap();
 		modelo.put("titulo", "Editar " + usuario.getNombreDeUsuario());
 		modelo.put("usuario", usuario); modelo.put("nombreUsuario", request.getSession().getAttribute("NOMBRE"));
-  
+		modelo.put("rol", request.getSession().getAttribute("ROL"));
+
 		return new ModelAndView("editarUsuario", modelo); 
 	}
 	  
 	@RequestMapping(path = "/validarEditarUsuario", method = RequestMethod.POST)
 	public ModelAndView validarEditarUsuario(@ModelAttribute("usuario") UsuarioModel usuario, HttpServletRequest request) { 
+		Long rol = (Long)request.getSession().getAttribute("ROL");
+		if (rol != Rol.ADMIN.getId()) 
+			return new ModelAndView ("redirect:/logout");
+		
 		ModelMap modelo = new ModelMap();
 		modelo.put("titulo", "Editar Usuario"); modelo.put("nombreUsuario", request.getSession().getAttribute("NOMBRE"));
-  
+		modelo.put("rol", request.getSession().getAttribute("ROL"));
+
 		usuarioService.modificarUsuario(usuario); 
 		return new ModelAndView("redirect:/usuarios");
   
@@ -98,9 +116,14 @@ import ar.edu.unlam.tallerweb1.servicios.UsuarioService;
   
 	@RequestMapping("/validarEliminarUsuario") 
 	public ModelAndView validarEliminarUsuario(@RequestParam("id") Long id, HttpServletRequest request) { 
+		Long rol = (Long)request.getSession().getAttribute("ROL");
+		if (rol != Rol.ADMIN.getId()) 
+			return new ModelAndView ("redirect:/logout");
+		
 		ModelMap modelo = new ModelMap();
 		modelo.put("nombreUsuario", request.getSession().getAttribute("NOMBRE"));
-  
+		modelo.put("rol", request.getSession().getAttribute("ROL"));
+
 		if (usuarioService.validarEliminarUsuario(id) == true) {
 			modelo.put("estadoEliminar", "El usuario se elimino exitosamente");
 			usuarioService.eliminarUsuarioPorId(id);
